removed data module
Joint up with lib.py
